pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "math-mentor"
version = "1.0.0"
description = "Multimodal math problem solver with multi-agent pipeline"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "easyocr>=1.7.0",
    "SpeechRecognition>=3.10.0",
    "sympy>=1.12",
    "streamlit>=1.28.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]
